requirements.txt
openai>=1.0.0
pyyaml>=6.0
pandas>=1.5.0
numpy>=1.24.0
python-dotenv>=1.0.0
jinja2>=3.1.0
rich>=13.0.0
click>=8.1.0